Refactorizando estructura de paquetes para el buen uso de la arquitectura Hexagonal
