+++ src/main/java/mrodriguezdev/me/apicorreo/application/EmailUseCase.java
@@ -1,20 +1,20 @@
-package mrodriguezdev.me.apicorreo.application.services;
+package mrodriguezdev.me.apicorreo.application;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
-import mrodriguezdev.me.apicorreo.domains.configuration.Apicorreo;
-import mrodriguezdev.me.apicorreo.domains.models.EmailDTO;
-import mrodriguezdev.me.apicorreo.domains.models.EmailResponse;
-import mrodriguezdev.me.apicorreo.domains.ports.in.EmailInputPort;
-import mrodriguezdev.me.apicorreo.domains.ports.out.EmailOutputPort;
+import mrodriguezdev.me.apicorreo.domain.configuration.Apicorreo;
+import mrodriguezdev.me.apicorreo.domain.model.email.EmailDTO;
+import mrodriguezdev.me.apicorreo.domain.model.email.EmailResponse;
+import mrodriguezdev.me.apicorreo.infraestructure.ports.in.EmailInputPort;
+import mrodriguezdev.me.apicorreo.infraestructure.ports.out.EmailOutputPort;
 
 import java.time.LocalDate;
 
 @ApplicationScoped
-public class EmailUserCase implements EmailInputPort {
+public class EmailUseCase implements EmailInputPort {
 
     @Inject
-    Apicorreo apicorreo;
+    Apicorreo configs;
 
     @Inject
     EmailOutputPort emailOutputPort;
@@ -31,7 +31,7 @@ public class EmailUserCase implements EmailInputPort {
         emailResponse.setCorreo(emailDTO.getCorreo());
         emailResponse.setAsunto(emailDTO.getAsunto());
         emailResponse.setMensaje(emailDTO.getMensaje());
-        emailResponse.setDomain(this.apicorreo.webAppDomain());
+        emailResponse.setDomain(this.configs.webApp().domain());
         emailResponse.setCurrentyear(LocalDate.now().getYear());
         return emailResponse;
     }
