+++ src/main/java/mrodriguezdev/me/apicorreo/domain/configuration/Apicorreo.java
@@ -1,13 +1,13 @@
-package mrodriguezdev.me.apicorreo.domains.configuration;
+package mrodriguezdev.me.apicorreo.domain.configuration;
 
 import io.smallrye.config.ConfigMapping;
 
 @ConfigMapping(prefix = "apicorreo")
 public interface Apicorreo {
-
     String personalEmail();
-
     String bussinesEmail();
-
-    String webAppDomain();
+    WebApp webApp();
+    interface WebApp {
+        String domain();
+    }
 }
