--- src/main/java/mrodriguezdev/me/apicorreo/domains/ports/in/EmailInputPort.java
@@ -1,9 +1,0 @@
-package mrodriguezdev.me.apicorreo.domains.ports.in;
-
-import jakarta.enterprise.context.ApplicationScoped;
-import mrodriguezdev.me.apicorreo.domains.models.EmailDTO;
-
-@ApplicationScoped
-public interface EmailInputPort {
-    void sendMail(EmailDTO emailDTO);
-}
