+++ src/main/java/mrodriguezdev/me/apicorreo/infraestructure/adapters/in/http/EmailResource.java
@@ -1,4 +1,4 @@
-package mrodriguezdev.me.apicorreo.infraestructure.controllers;
+package mrodriguezdev.me.apicorreo.infraestructure.adapters.in.http;
 
 import jakarta.inject.Inject;
 import jakarta.ws.rs.Consumes;
@@ -7,13 +7,13 @@ import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
-import mrodriguezdev.me.apicorreo.domains.models.EmailDTO;
-import mrodriguezdev.me.apicorreo.domains.ports.in.EmailInputPort;
+import mrodriguezdev.me.apicorreo.domain.model.email.EmailDTO;
+import mrodriguezdev.me.apicorreo.infraestructure.ports.in.EmailInputPort;
 
 @Path("email")
 @Produces(MediaType.APPLICATION_JSON)
 @Consumes(MediaType.APPLICATION_JSON)
-public class EmailController {
+public class EmailResource {
 
     @Inject
     EmailInputPort emailInputPort;
