--- src/main/java/mrodriguezdev/me/apicorreo/infraestructure/exceptions/InternalServerErrorException.java
@@ -1,17 +1,0 @@
-package mrodriguezdev.me.apicorreo.infraestructure.exceptions;
-
-import mrodriguezdev.me.apicorreo.domains.models.CustomExceptionResponse;
-
-public class InternalServerErrorException extends CustomException {
-    public InternalServerErrorException() {
-        super("Internal Server error", new CustomExceptionResponse<>("Internal Server error",500));
-    }
-
-    public InternalServerErrorException(String message) {
-        super(message, new CustomExceptionResponse<>(message, 500));
-    }
-
-    public InternalServerErrorException(CustomExceptionResponse<?> response) {
-        super(response);
-    }
-}
