+++ src/main/java/mrodriguezdev/me/apicorreo/infraestructure/ports/in/EmailInputPort.java
@@ -1,0 +1,7 @@
+package mrodriguezdev.me.apicorreo.infraestructure.ports.in;
+
+import mrodriguezdev.me.apicorreo.domain.model.email.EmailDTO;
+
+public interface EmailInputPort {
+    void sendMail(EmailDTO emailDTO);
+}
