+++ src/main/java/mrodriguezdev/me/apicorreo/infraestructure/ws/BadRequestException.java
@@ -1,0 +1,15 @@
+package mrodriguezdev.me.apicorreo.infraestructure.ws;
+
+public class BadRequestException extends CustomException {
+    public BadRequestException() {
+        super("Bad Request", new ResponseWS<>("Bad Request", 400));
+    }
+
+    public BadRequestException(String message) {
+        super(message, new ResponseWS<>(message, 400));
+    }
+
+    public BadRequestException(ResponseWS<?> response) {
+        super(response);
+    }
+}
