+++ src/main/java/mrodriguezdev/me/apicorreo/infraestructure/ws/InternalServerErrorException.java
@@ -1,0 +1,15 @@
+package mrodriguezdev.me.apicorreo.infraestructure.ws;
+
+public class InternalServerErrorException extends CustomException {
+    public InternalServerErrorException() {
+        super("Internal Server error", new ResponseWS<>("Internal Server error",500));
+    }
+
+    public InternalServerErrorException(String message) {
+        super(message, new ResponseWS<>(message, 500));
+    }
+
+    public InternalServerErrorException(ResponseWS<?> response) {
+        super(response);
+    }
+}
