+++ src/main/java/mrodriguezdev/me/apicorreo/infraestructure/ws/ResponseWS.java
@@ -1,29 +1,29 @@
-package mrodriguezdev.me.apicorreo.domains.models;
+package mrodriguezdev.me.apicorreo.infraestructure.ws;
 
 import io.quarkus.runtime.annotations.RegisterForReflection;
 
 @RegisterForReflection
-public class CustomExceptionResponse<T> {
+public class ResponseWS<T> {
     public T content;
     public String description;
     public String error;
     public int status;
 
-    public CustomExceptionResponse() {
+    public ResponseWS() {
     }
 
-    public CustomExceptionResponse(String description, String error, int status) {
+    public ResponseWS(String description, String error, int status) {
         this.description = description;
         this.error = error;
         this.status = status;
     }
 
-    public CustomExceptionResponse(T content, int status) {
+    public ResponseWS(T content, int status) {
         this.content = content;
         this.status = status;
     }
 
-    public CustomExceptionResponse(T content, String description, String error, int status) {
+    public ResponseWS(T content, String description, String error, int status) {
         this.content = content;
         this.description = description;
         this.error = error;
