+++ src/main/resources/application.properties
@@ -7,11 +7,6 @@ quarkus.smallrye-openapi.auto-add-tags=false
 mp.openapi.scan.disable=true
 quarkus.swagger-ui.path=/documentacion
 
-# ===== Generales =====
-apicorreo.personal-email=${MAIL_CC_ADDRES}
-apicorreo.bussines-email=${MAIL_ADDRES}
-apicorreo.web-app-domain=${WEB_APP_DOMAIN}
-
 # ===== Cors =====
 quarkus.http.cors=true
 quarkus.http.cors.origins=${CORS_ORIGIN}
@@ -26,3 +21,8 @@ quarkus.mailer.start-tls=REQUIRED
 quarkus.mailer.username=${MAIL_ADDRES}
 quarkus.mailer.password=${MAIL_PASSWORD}
 quarkus.mailer.mock=false
+
+# ===== Parametros apicorreo =====
+apicorreo.personal-email=${MAIL_CC_ADDRES}
+apicorreo.bussines-email=${MAIL_ADDRES}
+apicorreo.web-app.domain=${WEB_APP_DOMAIN}
