+++ src/main/resources/templates/correo_template.html
@@ -1,4 +1,4 @@
-{@mrodriguezdev.me.apicorreo.domains.models.EmailResponse email}
+{@mrodriguezdev.me.apicorreo.domain.model.email.EmailResponse email}
 <!DOCTYPE html>
 <html lang="es">
 <head>
